Reorganize C files to separate responsibilities between tracing and reporting
